BlackStone/requirements.txt
Flask ~= 2.0.1
grpcio-tools ~= 1.38.1
Jinja2 ~= 3.0.1
pytest ~= 6.2.4